Restructured and polished code.
