+++ BookyBook.Data/Data.Users.json
@@ -4,7 +4,7 @@
     "Email": "a",
     "Password": "a",
     "RegistrationDate": "2023-12-16T01:14:32.5976736+01:00",
-    "PenaltyFee": 6.25,
+    "PenaltyFee": 0,
     "IdNumber": 1111
   },
   {
