+++ BookyBook.Data/UserData.cs
@@ -7,7 +7,11 @@ namespace BookyBook.Data;
 public class UserData
 {
     public List<User>? UsersList = new();
-    public string UserJsonPath = @"..\BookyBook.Data\Data.Users.json";
+    private readonly string UserJsonPath = @"..\BookyBook.Data\Data.Users.json";
+    public UserData()
+    {
+        GetRegisteredUsers();
+    }
     public void AddUser(User user)
     {
         UsersList.Add(user);
