+++ BookyBook.Presentation/bin/Debug/net6.0/BookBook.Presentation.deps.json
@@ -44,7 +44,8 @@
       "BookyBook.Service/1.0.0": {
         "dependencies": {
           "BookyBook.Data": "1.0.0",
-          "BookyBook.Domain": "1.0.0"
+          "BookyBook.Domain": "1.0.0",
+          "Spectre.Console": "0.48.0"
         },
         "runtime": {
           "BookyBook.Service.dll": {}
