+++ BookyBook.Service/BookyBook.Service.csproj
@@ -5,6 +5,10 @@
     <ProjectReference Include="..\BookyBook.Data\BookyBook.Data.csproj" />
   </ItemGroup>
 
+  <ItemGroup>
+    <PackageReference Include="Spectre.Console" Version="0.48.0" />
+  </ItemGroup>
+
   <PropertyGroup>
     <TargetFramework>net6.0</TargetFramework>
     <ImplicitUsings>enable</ImplicitUsings>
