+++ BookyBook.Service/bin/Debug/net6.0/BookyBook.Service.deps.json
@@ -9,12 +9,25 @@
       "BookyBook.Service/1.0.0": {
         "dependencies": {
           "BookyBook.Data": "1.0.0",
-          "BookyBook.Domain": "1.0.0"
+          "BookyBook.Domain": "1.0.0",
+          "Spectre.Console": "0.48.0"
         },
         "runtime": {
           "BookyBook.Service.dll": {}
         }
       },
+      "Spectre.Console/0.48.0": {
+        "dependencies": {
+          "System.Memory": "4.5.5"
+        },
+        "runtime": {
+          "lib/net6.0/Spectre.Console.dll": {
+            "assemblyVersion": "0.0.0.0",
+            "fileVersion": "0.48.0.0"
+          }
+        }
+      },
+      "System.Memory/4.5.5": {},
       "BookyBook.Data/1.0.0": {
         "dependencies": {
           "BookyBook.Domain": "1.0.0"
@@ -35,6 +48,20 @@
       "type": "project",
       "serviceable": false,
       "sha512": ""
+    },
+    "Spectre.Console/0.48.0": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-4Mc1UT7Azgtyb8FyNwK5FZmoZbKuT5PmY7ZwaKUytjD5kGFMNBACpOZTwYtkuY377YkYtZYBeDDTJUwTW86QXw==",
+      "path": "spectre.console/0.48.0",
+      "hashPath": "spectre.console.0.48.0.nupkg.sha512"
+    },
+    "System.Memory/4.5.5": {
+      "type": "package",
+      "serviceable": true,
+      "sha512": "sha512-XIWiDvKPXaTveaB7HVganDlOCRoj03l+jrwNvcge/t8vhGYKvqV+dMv6G4SAX2NoNmN0wZfVPTAlFwZcZvVOUw==",
+      "path": "system.memory/4.5.5",
+      "hashPath": "system.memory.4.5.5.nupkg.sha512"
     },
     "BookyBook.Data/1.0.0": {
       "type": "project",
